# openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "Task API", "version": "1.0.0", "description": "A simple in-memory CRUD API for managing to-do tasks." },
  "servers": [{ "url": "http://localhost:3000" }],
  "paths": {
    "/": { "get": { "summary": "Describe the API", "responses": { "200": { "description": "API metadata returned" } } } },
    "/health": { "get": { "summary": "Health check", "responses": { "200": { "description": "Server is healthy" } } } },
    "/tasks": {
      "get": { "summary": "List all tasks", "responses": { "200": { "description": "Task list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Task" } } } } } } },
      "post": { "summary": "Create a task", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateTask" }, "example": { "title": "Buy milk" } } } }, "responses": { "201": { "description": "Task created" }, "400": { "description": "Invalid request" } } }
    },
    "/tasks/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
      "get": { "summary": "Get one task", "responses": { "200": { "description": "Task found" }, "404": { "description": "Task not found" } } },
      "put": { "summary": "Update a task", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateTask" }, "examples": { "mark_done": { "value": { "done": true } }, "rename": { "value": { "title": "Buy oat milk" } } } } } }, "responses": { "200": { "description": "Task updated" }, "400": { "description": "Invalid request" }, "404": { "description": "Task not found" } } },
      "delete": { "summary": "Delete a task", "responses": { "204": { "description": "Task deleted" }, "404": { "description": "Task not found" } } }
    }
  },
  "components": {
    "schemas": {
      "Task": { "type": "object", "required": ["id", "title", "done"], "properties": { "id": { "type": "integer", "example": 1 }, "title": { "type": "string", "example": "Buy milk" }, "done": { "type": "boolean", "example": false } } },
      "CreateTask": { "type": "object", "required": ["title"], "properties": { "title": { "type": "string", "example": "Buy milk" } } },
      "UpdateTask": { "type": "object", "minProperties": 1, "properties": { "title": { "type": "string", "example": "Buy oat milk" }, "done": { "type": "boolean", "example": true } } }
    }
  }
}